registerAgent/requirements.txt
langchain>=0.1.0
langchain-aws>=0.1.0
langchain-core>=0.1.0
langgraph>=0.1.0
boto3>=1.34.0
python-dotenv>=1.0.0
bedrock-agentcore>=1.0.5
aws-opentelemetry-distro>=0.10.0
requests>=2.31.0
strands-agents>=0.1.0
strands-agents-tools>=0.1.0
openfga-sdk>=0.3.5
aiohttp>=3.9.0
